+++ packages/nhsuk-frontend/src/nhsuk/components/back-link/macro-options.mjs
@@ -9,53 +9,62 @@ const options = {
   id: {
     type: 'string',
     required: false,
-    description: 'The ID of the back link.'
+    description: 'The ID of the back link.',
+    released: '10.0.0'
   },
   text: {
     type: 'string',
     required: false,
     description:
-      'Text to use within the back link component. If `html` is provided, the `text` argument will be ignored. Defaults to `"Back"`.'
+      'Text to use within the back link component. If `html` is provided, the `text` argument will be ignored. Defaults to `"Back"`.',
+    released: '1.0.0'
   },
   html: {
     type: 'string',
     required: false,
     description:
-      'HTML to use within the back link component. If `html` is provided, the `text` argument will be ignored. Defaults to `"Back"`.'
+      'HTML to use within the back link component. If `html` is provided, the `text` argument will be ignored. Defaults to `"Back"`.',
+    released: '1.0.0'
   },
   href: {
     type: 'string',
     required: false,
-    description: 'The value of the link `href` attribute.'
+    description: 'The value of the link `href` attribute.',
+    released: '1.0.0'
   },
   variant: {
     type: 'string',
     required: false,
     description:
-      'Optional variant of back link. You can use only `"reverse"` or empty values with this option.'
+      'Optional variant of back link. You can use only `"reverse"` or empty values with this option.',
+    released: '10.4.0'
   },
   classes: {
     type: 'string',
     required: false,
-    description: 'Classes to add to the container.'
+    description: 'Classes to add to the container.',
+    released: '1.0.0'
   },
   attributes: {
     type: 'object',
     required: false,
     description:
-      'HTML attributes (for example data attributes) to add to the tag.'
+      'HTML attributes (for example data attributes) to add to the tag.',
+    released: '1.0.0'
   },
   visuallyHiddenText: {
     type: 'string',
     required: false,
     description:
-      'An optional visually hidden prefix used before the back link text, for example `"Back to"` used by the breadcrumbs component.'
+      'An optional visually hidden prefix used before the back link text, for example `"Back to"` used by the breadcrumbs component.',
+    released: '10.1.0'
   },
   element: {
     type: 'string',
     required: false,
     description:
-      'HTML element for the back link component – `"button"` or `"a"`. If not provided, defaults to `"a"`.'
+      'HTML element for the back link component – `"button"` or `"a"`. If not provided, defaults to `"a"`.',
+    released: '6.2.0'
   }
 }