+++ packages/nhsuk-frontend/src/nhsuk/components/label/macro-options.mjs
@@ -9,52 +9,61 @@ const options = {
   id: {
     type: 'string',
     required: false,
-    description: 'The ID of the label.'
+    description: 'The ID of the label.',
+    released: '10.0.0'
   },
   text: {
     type: 'string',
     required: true,
     description:
-      'If `html` is set, this is not required. Text to use within the label. If `html` is provided, the `text` option will be ignored.'
+      'If `html` is set, this is not required. Text to use within the label. If `html` is provided, the `text` option will be ignored.',
+    released: '1.0.0'
   },
   html: {
     type: 'string',
     required: true,
     description:
-      'If `text` is set, this is not required. HTML to use within the label. If `html` is provided, the `text` option will be ignored.'
+      'If `text` is set, this is not required. HTML to use within the label. If `html` is provided, the `text` option will be ignored.',
+    released: '1.0.0'
   },
   caller: {
     type: 'nunjucks-block',
     required: false,
     description:
-      'Not strictly a parameter but Nunjucks code convention. Using a `call` block enables you to call a macro with all the text inside the tag. This is helpful if you want to pass a lot of content into a macro. To use it, you will need to wrap the entire label component in a `call` block.'
+      'Not strictly a parameter but Nunjucks code convention. Using a `call` block enables you to call a macro with all the text inside the tag. This is helpful if you want to pass a lot of content into a macro. To use it, you will need to wrap the entire label component in a `call` block.',
+    released: '10.2.0'
   },
   for: {
     type: 'string',
     required: false,
     description:
-      'The value of the `for` attribute, the ID of the input the label is associated with.'
+      'The value of the `for` attribute, the ID of the input the label is associated with.',
+    released: '1.0.0'
   },
   isPageHeading: {
     type: 'boolean',
     required: false,
-    description: 'Whether the label also acts as the heading for the page.'
+    description: 'Whether the label also acts as the heading for the page.',
+    released: '1.0.0'
   },
   size: {
     type: 'string',
     required: false,
-    description: 'Size of the label – `"s"`, `"m"`, `"l"` or `"xl"`.'
+    description: 'Size of the label – `"s"`, `"m"`, `"l"` or `"xl"`.',
+    released: '10.2.0'
   },
   classes: {
     type: 'string',
     required: false,
-    description: 'Classes to add to the label tag.'
+    description: 'Classes to add to the label tag.',
+    released: '1.0.0'
   },
   attributes: {
     type: 'object',
     required: false,
     description:
-      'HTML attributes (for example data attributes) to add to the label tag.'
+      'HTML attributes (for example data attributes) to add to the label tag.',
+    released: '1.0.0'
   }
 }